With these changes, Issue #42 COULD be closed.
While the previous 'consolidation' changes in the Issue42 branch
should be fairly uncontroversial (and therefore have been merged
into development by me), the changes in this commit work (tested
with Mode1, Mode3, and even Mode2), but kind of revert the files

   GridMain/AMR/Amrex/block_1lev_iterator.F90
   GridMain/AMR/Amrex/gr_iterator.F90

back to a state after copying from AmrexTransition versions but
before much cleanup by Jared (and then apply some changes on top
of that). Sorry, I did not have the time or patience to carefully
go through all the diffs and apply individually.