(Issue #217) Towards a working unitTest for 2d-cyl gravity
Adapted some simulation-specific Driver routine implementations so
that they can be compiled in FLASH5. Testing with
```
  ./setup unitTest/Gravity/Poisson3 -2d +cylindrical -auto -maxblocks=600 -debug +noio +pm4dev +newMpole -parfile=flash_2dcyl.par -objdir=UnitTest_Gravity_Poisson3_2d -makefile=gnu
```

Compilation now fails in Gravity_potentialListOfBlocks.F90.

There must be a version of this file that has already been adapted
for use with FLASH5, but it does not appear to be in the master branch.
I assume Saurabh (@chaw0023) knows where such a version is.

This should be updated anyway as soon as possible to work with the
newer convention, in which Gravity_potentialListOfBlocks is renamed
to just Gravity_potential. The change of the Gravity unit to that newer convention
is in the Multipole-issue133 branch, a pull request (#207) is pending to get
it into master, waiting for @jared321 to merge.