include all partial derivatives in the count on 2nd line
The 2nd number on the 2nd line should count all partial derivatives.
That way, the reader can know in a simple way how many numbers to
read leater.

If for some reason it would be beneficial to directly store something
like "highest derivative in any direction" close to the beginning
of the file (which appears to be the meaning of the 2nd number
before this correction), the format could be enhanced to add such
a number, maybe at the end of the 2nd line.